- tag more functions with gnu format attribute - fix format string bugs. still some left

+++ code/game/g_session.c
@@ -54,7 +54,7 @@ void G_WriteClientSessionData( gclient_t *client ) {
 		client->sess.teamLeader
 		);
 
-	var = va( "session%i", client - level.clients );
+	var = va( "session%i", (int)(client - level.clients) );
 
 	trap_Cvar_Set( var, s );
 }
@@ -75,7 +75,7 @@ void G_ReadSessionData( gclient_t *client ) {
 	int spectatorState;
 	int sessionTeam;
 
-	var = va( "session%i", client - level.clients );
+	var = va( "session%i", (int)(client - level.clients) );
 	trap_Cvar_VariableStringBuffer( var, s, sizeof(s) );
 
 	sscanf( s, "%i %i %i %i %i %i %i",
